requirements.txt
amqp==5.1.1
asgiref==3.6.0
attrs==22.2.0
billiard==3.6.4.0
blinker==1.6.2
Brotli==1.0.9
celery==5.2.7
certifi==2023.5.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.33.0
colorama==0.4.6
dj-database-url==1.2.0
Django==4.1.7
django-celery-results==2.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.0
Flask==2.3.2
gunicorn==20.1.0
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
kombu==5.2.4
MarkupSafe==2.1.3
phonenumbers==8.13.14
Pillow==9.5.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
pycodestyle==2.10.0
PyJWT==2.6.0
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.16
validate-docbr==1.10.0
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.3.6
whitenoise==6.4.0
